requirements.txt
streamlit
pytesseract
pillow
pandas
plotly
python-dateutil
st-gsheets-connection
requests